format last component/forms & utils, convert _doc, index and sitemap
