+++ pages/index.tsx
@@ -1,118 +1,66 @@
-import Head from "next/head";
-import Image from "next/image";
-import localFont from "next/font/local";
-import styles from "@/styles/Home.module.css";
+import React from 'react';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+import { faGithub, faLinkedin } from '@fortawesome/free-brands-svg-icons';
+import { faEnvelope } from '@fortawesome/free-solid-svg-icons';
+import Hero from '../components/sections/Hero';
+import AboutSection from '../components/sections/AboutSection';
+import ProjectsSection from '../components/sections/ProjectsSection';
+import BreakSection from '../components/sections/BreakSection';
+import classes from '../styles/index.module.css';
+import LinkWrapper from '../components/utils/LinkWrapper';
+import PageHead from '../components/layout/PageHead';
+import dbConnect from '../utils/dbConnect';
+import Project from '../models/Project';
+import { Project as ProjectType } from '../interfaces/project';
 
-const geistSans = localFont({
-  src: "./fonts/GeistVF.woff",
-  variable: "--font-geist-sans",
-  weight: "100 900",
-});
-const geistMono = localFont({
-  src: "./fonts/GeistMonoVF.woff",
-  variable: "--font-geist-mono",
-  weight: "100 900",
-});
+interface Props {
+  projects: ProjectType[];
+}
 
-export default function Home() {
+export default function Home({ projects }: Props) {
   return (
-    <>
-      <Head>
-        <title>Create Next App</title>
-        <meta name="description" content="Generated by create next app" />
-        <meta name="viewport" content="width=device-width, initial-scale=1" />
-        <link rel="icon" href="/favicon.ico" />
-      </Head>
-      <div
-        className={`${styles.page} ${geistSans.variable} ${geistMono.variable}`}
-      >
-        <main className={styles.main}>
-          <Image
-            className={styles.logo}
-            src="/next.svg"
-            alt="Next.js logo"
-            width={180}
-            height={38}
-            priority
-          />
-          <ol>
-            <li>
-              Get started by editing <code>pages/index.tsx</code>.
-            </li>
-            <li>Save and see your changes instantly.</li>
-          </ol>
+    <div>
+      <PageHead title="James McGahn | Home" />
+      <Hero h1="" h2="" />
 
-          <div className={styles.ctas}>
+      <AboutSection />
+      <BreakSection color="" url="/img/breaker6.jpg">
+        <div className={classes.iconSection}>
+          <div className={classes.icon}>
             <a
-              className={styles.primary}
-              href="https://vercel.com/new?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
+              href="https://github.com/JamesMcGahn"
               target="_blank"
               rel="noopener noreferrer"
             >
-              <Image
-                className={styles.logo}
-                src="/vercel.svg"
-                alt="Vercel logomark"
-                width={20}
-                height={20}
-              />
-              Deploy now
-            </a>
-            <a
-              href="https://nextjs.org/docs?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
-              target="_blank"
-              rel="noopener noreferrer"
-              className={styles.secondary}
-            >
-              Read our docs
+              <FontAwesomeIcon icon={faGithub} />
             </a>
           </div>
-        </main>
-        <footer className={styles.footer}>
-          <a
-            href="https://nextjs.org/learn?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            <Image
-              aria-hidden
-              src="/file.svg"
-              alt="File icon"
-              width={16}
-              height={16}
-            />
-            Learn
-          </a>
-          <a
-            href="https://vercel.com/templates?framework=next.js&utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            <Image
-              aria-hidden
-              src="/window.svg"
-              alt="Window icon"
-              width={16}
-              height={16}
-            />
-            Examples
-          </a>
-          <a
-            href="https://nextjs.org?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            <Image
-              aria-hidden
-              src="/globe.svg"
-              alt="Globe icon"
-              width={16}
-              height={16}
-            />
-            Go to nextjs.org →
-          </a>
-        </footer>
-      </div>
-    </>
+          <div className={classes.icon}>
+            <a
+              href="https://www.linkedin.com/in/james-mcgahn-579067156/"
+              target="_blank"
+              rel="noopener noreferrer"
+            >
+              <FontAwesomeIcon icon={faLinkedin} />
+            </a>
+          </div>
+          <div className={classes.icon}>
+            <LinkWrapper to="/contact">
+              <FontAwesomeIcon icon={faEnvelope} />
+            </LinkWrapper>
+          </div>
+        </div>
+      </BreakSection>
+      <ProjectsSection projects={projects} mainPage />
+    </div>
   );
+}
+
+export async function getStaticProps() {
+  await dbConnect();
+  const projects = await Project.find({ mainPage: true }).lean();
+  return {
+    props: { projects: JSON.parse(JSON.stringify(projects)) },
+    revalidate: 3600,
+  };
 }
