+++ utils/dbConnect.ts
@@ -1,17 +1,16 @@
 import mongoose from 'mongoose';
 
 async function dbConnect() {
-    if (mongoose.connection.readyState === 1) {
-        return
+  if (mongoose.connection.readyState === 1) {
+    return;
+  }
+  try {
+    if (process.env.MONGO_URI) {
+      await mongoose.connect(process.env.MONGO_URI);
     }
-    try {
-        if (process.env.MONGO_URI) {
-            await mongoose.connect(process.env.MONGO_URI)
-        }
-
-    } catch (error) {
-        console.error(error)
-    }
+  } catch (error) {
+    console.error(error);
+  }
 }
 
-export default dbConnect
+export default dbConnect;
